Moved all written implementation to controller package
